Use shared ActiveSupport integration in Sidekiq and Rails integrations
